requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
anyio==3.6.2
    # via
    #   dagger-io
    #   httpcore
attrs==22.2.0
    # via
    #   cattrs
    #   dagger-io
backoff==2.2.1
    # via gql
cattrs==22.2.0
    # via dagger-io
certifi==2022.12.7
    # via
    #   httpcore
    #   httpx
click==8.1.3
    # via typer
colorama==0.4.6
    # via typer
commonmark==0.9.1
    # via rich
dagger-io==0.2.1
    # via -r requirements.in
exceptiongroup==1.1.0
    # via cattrs
gql==3.4.0
    # via dagger-io
graphql-core==3.2.3
    # via gql
h11==0.14.0
    # via httpcore
httpcore==0.16.3
    # via httpx
httpx==0.23.1
    # via dagger-io
idna==3.4
    # via
    #   anyio
    #   rfc3986
    #   yarl
multidict==6.0.4
    # via yarl
pygments==2.13.0
    # via rich
rfc3986[idna2008]==1.5.0
    # via httpx
rich==12.6.0
    # via
    #   -r requirements.in
    #   typer
shellingham==1.5.0
    # via typer
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
typer[all]==0.7.0
    # via dagger-io
yarl==1.8.2
    # via gql
